+++ cmd/oh/style/style.go
@@ -70,7 +70,6 @@ var (
 	History        Style = hex(mauve)
 	PendingPrefix  Style = col.Underline
 	ScrolledInput  Style = Dim
-	DisabledInput  Style = Dim
 	ChosenRow      Style = hex(copper)
 	LowPrice       Style = hex(lime)
 	MediumPrice    Style = hex(steel)